+++ src/Controller/CommentController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controller\User;
+namespace App\Controller;
 
 use App\Entity\Comment;
 use App\Repository\CommentRepository;
@@ -10,7 +10,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
-class UserCommentController extends AbstractController
+class CommentController extends AbstractController
 {
     /**
      * @var EntityManagerInterface
